Reusing some common methods of drive_enclosure, sas_interconnect and interconnect
