+++ libraries/resource_providers/api200.rb
@@ -24,5 +24,7 @@ module OneviewCookbook
   end
 end
 
+# Load the helper files:
+Dir[File.dirname(__FILE__) + '/helpers/*.rb'].each { |file| require file }
 # Load all API-specific resources:
 Dir[File.dirname(__FILE__) + '/api200/*.rb'].each { |file| require file }
